Add callerProcessId parameter to COM launcher methods
I must have the handle of the process that requested the
process be launched so that I can return the spawned process'
I/O handles to it in a way such that the handles will be usable
once they cross the process boundary. To do this, I must know
the process ID of the process calling the method.